AXON-533: refactor e2e test file structure (#507)
* AXON-533: refactor e2e test file structure

* AXON-533: use one worker for tests as it was in previous setup
